requirements.txt
pandas
numpy
Pillow
opencv-python-headless
torch
torchvision